Module yup_oauth2::authenticator

Module containing the core functionality for OAuth2 Authentication.

Structs

AccessTokenAuthenticator

Create a access token authenticator for use with pre-generated access tokens

ApplicationDefaultCredentialsAuthenticator

Create an authenticator that uses a application default credentials.

Authenticator

Authenticator is responsible for fetching tokens, handling refreshing tokens, and optionally persisting tokens to disk.

AuthenticatorBuilder

Configure an Authenticator using the builder pattern.

AuthorizedUserAuthenticator

Create an authenticator that uses an authorized user credentials.

DefaultHyperClient

The builder value used when the default hyper client should be used.

DeviceFlowAuthenticator

Create an authenticator that uses the device flow.

InstalledFlowAuthenticator

Create an authenticator that uses the installed flow.

ServiceAccountAuthenticator

Create an authenticator that uses a service account.

ServiceAccountImpersonationAuthenticator

Create a access token authenticator that uses user secrets to impersonate a service account.

Enums

ApplicationDefaultCredentialsTypes

Types of authenticators provided by ApplicationDefaultCredentialsAuthenticator

Traits

HyperClientBuilder

A trait implemented for any hyper::Client as well as the DefaultHyperClient.

Type Aliases

DefaultAuthenticator

Default authenticator type