Module yup_oauth2::authenticator_delegate

Module containing types related to delegates.

Structs

DefaultDeviceFlowDelegate
Uses all default implementations in the DeviceFlowDelegate trait.

DefaultInstalledFlowDelegate
Uses all default implementations in the DeviceFlowDelegate trait.

DeviceAuthResponse
Contains state of pending authentication requests

Traits

DeviceFlowDelegate
DeviceFlowDelegate methods are called when a device flow needs to ask the application what to do in certain cases.

InstalledFlowDelegate
InstalledFlowDelegate methods are called when an installed flow needs to ask the application what to do in certain cases.