Trait yup_oauth2::authenticator_delegate::InstalledFlowDelegate

pub trait InstalledFlowDelegate: Send + Sync {
    fn redirect_uri(&self) -> Option<&str> { ... }

    fn present_user_url<'a>(
        &'a self, 
        url: &'a str, 
        need_code: bool
    ) -> Pin<Box<dyn Future<Output = Result<String, String>> + Send + 'a>> { ... }
}

InstalledFlowDelegate methods are called when an installed flow needs to ask the application what to do in certain cases.

Provided methods

fn redirect_uri(&self) -> Option<&str>

Configure a custom redirect uri if needed.

fn present_user_url<'a>(
&'a self,
url: &'a str,
need_code: bool
) -> Pin<Box<dyn Future<Output = Result<String, String>> + Send + 'a>>

We need the user to navigate to a URL using their browser and potentially paste back a code (or maybe not). Whether they have to enter a code depends on the InstalledFlowReturnMethod used.

Loading content...