Enum yup_oauth2::FlowType

pub enum FlowType {
    Device,
    InstalledInteractive,
    InstalledRedirect(u32),
}

All known authentication types, for suitable constants

Variants
Device

device authentication. Only works for certain scopes.

InstalledInteractive

installed app flow. Required for Drive, Calendar, Gmail...; Requires user to paste a code from the browser.

InstalledRedirect(u32)

Same as InstalledInteractive, but uses a redirect: The OAuth provider redirects the user's browser to a web server that is running on localhost. This may not work as well with the Windows Firewall, but is more comfortable otherwise. The integer describes which port to bind to (default: 8080)