Enum yup_oauth2::InstalledFlowReturnMethod

pub enum InstalledFlowReturnMethod {
    Interactive,
    HTTPRedirect(u32),
}

cf. https://developers.google.com/identity/protocols/OAuth2InstalledApp#choosingredirecturi

Variants

Interactive

Involves showing a URL to the user and asking to copy a code from their browser (default)

HTTPRedirect

Involves spinning up a local HTTP server and Google redirecting the browser to the server with a URL containing the code (preferred, but not as reliable). The parameter is the port to listen on.