Enum yup-oauth2::RefreshResult

pub enum RefreshResult {
    Error(HttpError),
    Refused(String),
    Success(Token),
}

All possible outcomes of the refresh flow

Variants

Error

Indicates connection failure

Refused

The server did not answer with a new token, providing the server message

Success

The refresh operation finished successfully, providing a new Token