Struct yup_oauth2::Authenticator

pub struct Authenticator<D, S, C> {
    // some fields omitted
}

A generalized authenticator which will keep tokens valid and store them.

It is the go-to helper to deal with any kind of supported authentication flow, which will be kept valid and usable.

Device Flow

This involves polling the authentication server in the given intervals until there is a definitive result.

These results will be passed the DeviceFlowHelperDelegate implementation to deal with * presenting the user code * inform the user about the progress or errors * abort the operation

Usage

Please have a look at the library's landing page.

Methods

impl<D, S, C> Authenticator<D, S, C> where D: AuthenticatorDelegate, S: TokenStorage, C: BorrowMut<Client>

fn new(secret: &ApplicationSecret, delegate: D, client: C, storage: S, flow_type: Option<FlowType>) -> Authenticator<D, S, C>

Returns a new Authenticator instance

Arguments

secret - usually obtained from a client secret file produced by the developer console

delegate - Used to further refine the flow of the authentication.

client - used for all authentication https requests

storage - used to cache authorization tokens tokens permanently. However, the implementation doesn't have any particular semantic requirement, which is why NullStorage and MemoryStorage can be used as well.

flow_type - the kind of authentication to use to obtain a token for the required scopes. If unset, it will be derived from the secret.