Struct yup_oauth2::ConsoleApplicationSecret

pub struct ConsoleApplicationSecret {
    pub web: Option<ApplicationSecret>,
    pub installed: Option<ApplicationSecret>,
}

A type to facilitate reading and writing the json secret file as returned by the google developer console

Fields
web: Option<ApplicationSecret> installed: Option<ApplicationSecret>