Struct yup-oauth2::DeviceFlow

pub struct DeviceFlow<C, NC> {
    // some fields omitted
}

Implements the Oauth2 Device Flow It operates in two steps: * obtain a code to show to the user * (repeatedly) poll for the user to authenticate your application

Methods

impl<C, NC> DeviceFlow<C, NC> where C: BorrowMut<Client<NC>>, NC: NetworkConnector

fn new(client: C) -> DeviceFlow<C, NC>

Examples

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
use oauth2::DeviceFlow;
 
let mut f = DeviceFlow::new(hyper::Client::new());

fn request_code<'b, T, I>(&mut self, client_id: &str, client_secret: &str, scopes: I) -> RequestResult where T: Str, I: IntoIterator<Item=&'b T>

The first step involves asking the server for a code that the user can type into a field at a specified URL. It is called only once, assuming there was no connection error. Otherwise, it may be called again until the state changes to PollResult::AuthorizationPending.

Arguments

client_id & client_secret - as obtained when registering your application

scopes - an iterator yielding String-like objects which are URLs defining what your application is able to do. It is considered good behaviour to authenticate only once, with all scopes you will ever require. However, you can also manage multiple tokens for different scopes, if your application is providing distinct read-only and write modes. # Handling the PollResult

will panic if called while our state is not PollResult::Error or PollResult::NeedToken # Examples See test-cases in source code for a more complete example.

fn poll_token(&mut self) -> PollResult

If the first call is successful, which is expected unless there is a network problem, the returned PollResult::AuthorizationPending variant contains enough information to poll within a given interval to at some point obtain a result which is not PollResult::AuthorizationPending.

Handling the PollResult

call within PollResult::AuthorizationPending.interval until the variant changes. Keep calling as desired, even after PollResult::Error.

Do not call after PollResult::Expired, PollResult::AccessDenied or PollResult::AccessGranted as the flow will do nothing anymore. Thus in any unsuccessful case, you will have to start over the entire flow, which requires a new instance of this type.

⚠️ Warning: We assume the caller doesn't call faster than interval and are not protected against this kind of mis-use. The latter will be indicated in PollResult::AuthorizationPending.server_message

Examples

See test-cases in source code for a more complete example.