Struct yup_oauth2::InstalledFlow

pub struct InstalledFlow<C> {
    // some fields omitted
}

Methods

impl<C> InstalledFlow<C> where C: BorrowMut<Client>

fn new(client: C, method: Option<InstalledFlowReturnMethod>) -> InstalledFlow<C>

Starts a new Installed App auth flow. If HTTPRedirect is chosen as method and the server can't be started, the flow falls back to Interactive.

fn obtain_token<'a, AD: AuthenticatorDelegate, S, T>(&mut self, auth_delegate: &mut AD, appsecret: &ApplicationSecret, scopes: S) -> Result<Token, Box<Error>> where T: AsRef<str> + 'a, S: Iterator<Item=&'a T>

Handles the token request flow; it consists of the following steps: . Obtain a auhorization code with user cooperation or internal redirect. . Obtain a token and refresh token using that code. . Return that token

It's recommended not to use the DefaultAuthenticatorDelegate, but a specialized one.