Struct yup_oauth2::PollInformation

pub struct PollInformation {
    pub user_code: String,
    pub verification_url: String,
    pub expires_at: DateTime<UTC>,
    pub interval: Duration,
}

Contains state of pending authentication requests

Fields

user_code

Code the user must enter ...

verification_url

... at the verification URL

expires_at

The user_code expires at the given time It's the time the user has left to authenticate your application

interval

The interval in which we may poll for a status change The server responds with errors of we poll too fast.