Struct yup_oauth2::Token

pub struct Token {
    pub access_token: String,
    pub refresh_token: String,
    pub token_type: String,
    pub expires_in: Option<i64>,
    pub expires_in_timestamp: Option<i64>,
}

Represents a token as returned by OAuth2 servers.

It is produced by all authentication flows. It authenticates certain operations, and must be refreshed once it reached it's expiry date.

The type is tuned to be suitable for direct de-serialization from server replies, as well as for serialization for later reuse. This is the reason for the two fields dealing with expiry - once in relative in and once in absolute terms.

Utility methods make common queries easier, see expired().

Fields

access_token

used when authenticating calls to oauth2 enabled services.

refresh_token

used to refresh an expired access_token.

token_type

The token type as string - usually 'Bearer'.

expires_in

access_token will expire after this amount of time. Prefer using expiry_date()

expires_in_timestamp

timestamp is seconds since epoch indicating when the token will expire in absolute terms. use expiry_date() to convert to DateTime.

Methods

impl Token

fn expired(&self) -> bool

Returns true if we are expired.

Panics

if our access_token is unset

fn expiry_date(&self) -> DateTime<UTC>

Returns a DateTime object representing our expiry date.

fn set_expiry_absolute(&mut self) -> &mut Token

Adjust our stored expiry format to be absolute, using the current time.