Trait yup-oauth2::AuthenticatorDelegate

pub trait AuthenticatorDelegate {
    fn connection_error(&mut self, HttpError) -> Retry { ... }
    fn request_failure(&mut self, RequestResult) { ... }
    fn expired(&mut self, DateTime<UTC>) { ... }
    fn denied(&mut self) { ... }
    fn pending(&mut self, &PollInformation) -> Retry { ... }
    fn present_user_code(&mut self, pi: PollInformation) { ... }
}

A partially implemented trait to interact with the Authenticator

The only method that needs to be implemented manually is present_user_code(...), as no assumptions are made on how this presentation should happen.

Provided Methods

fn connection_error(&mut self, HttpError) -> Retry

Called whenever there is an HttpError, usually if there are network problems.

Return retry information.

fn request_failure(&mut self, RequestResult)

The server denied the attempt to obtain a request code

fn expired(&mut self, DateTime<UTC>)

Called if the request code is expired. You will have to start over in this case. This will be the last call the delegate receives.

fn denied(&mut self)

Called if the user denied access. You would have to start over. This will be the last call the delegate receives.

fn pending(&mut self, &PollInformation) -> Retry

Called as long as we are waiting for the user to authorize us. Can be used to print progress information, or decide to time-out.

If the returned Retry variant is a duration.

Notes

Only used in DeviceFlow. Return value will only be used if it is larger than the interval desired by the server.

fn present_user_code(&mut self, pi: PollInformation)

The server has returned a user_code which must be shown to the user, along with the verification_url.

Notes

Will be called exactly once, provided we didn't abort during request_code phase.

Will only be called if the Authenticator's flow_type is FlowType::Device.