Crate yup_oauth2

This library can be used to acquire oauth2.0 authentication for services.

For your application to use this library, you will have to obtain an application id and secret by following this guide (for Google services) respectively the documentation of the API provider you want to connect to.

Device Flow Usage

As the DeviceFlow involves polling, the DeviceFlowHelper should be used as means to adhere to the protocol, and remain resilient to all kinds of errors that can occour on the way.

Service account "flow"

When using service account credentials, no user interaction is required. The access token can be obtained automatically using the private key of the client (which you can download from the API provider). See examples/service_account/ for an example on how to use service account credentials. See developers.google.com for a detailed description of the protocol. This crate implements OAuth for Service Accounts based on the Google APIs; it may or may not work with other providers.

Installed Flow Usage

The InstalledFlow involves showing a URL to the user (or opening it in a browser) and then either prompting the user to enter a displayed code, or make the authorizing website redirect to a web server spun up by this library and running on localhost.

In order to use the interactive method, use the InstalledInteractive FlowType; for the redirect method, use InstalledRedirect, with the port number to let the server listen on.

You can implement your own AuthenticatorDelegate in order to customize the flow; the InstalledFlow uses the present_user_url method.

The returned Token is stored permanently in the given token storage in order to authorize future API requests to the same scopes.

#[macro_use]
extern crate serde_derive;
 
use yup_oauth2::{Authenticator, DefaultAuthenticatorDelegate, PollInformation, ConsoleApplicationSecret, MemoryStorage, GetToken};
use serde_json as json;
use std::default::Default;
use hyper::{Client, net::HttpsConnector};
use hyper_native_tls::NativeTlsClient;

let secret = json::from_str::<ConsoleApplicationSecret>(SECRET).unwrap().installed.unwrap();
let res = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                        Client::with_connector(HttpsConnector::new(NativeTlsClient::new().unwrap())),
                        <MemoryStorage as Default>::default(), None)
                        .token(&["https://www.googleapis.com/auth/youtube.upload"]);
match res {
    Ok(t) => {
    // now you can use t.access_token to authenticate API calls within your
    // given scopes. It will not be valid forever, but Authenticator will automatically
    // refresh the token for you.
    },
    Err(err) => println!("Failed to acquire token: {}", err),
}

Structs

ApplicationSecret
Represents either 'installed' or 'web' applications in a json secrets file. See ConsoleApplicationSecret for more information

Authenticator
A generalized authenticator which will keep tokens valid and store them.

ConsoleApplicationSecret
A type to facilitate reading and writing the json secret file as returned by the google developer console

DefaultAuthenticatorDelegate
Uses all default implementations by AuthenticatorDelegate, and makes the trait's implementation usable in the first place.

DeviceFlow
Implements the Oauth2 Device Flow It operates in two steps:

DiskTokenStorage
Serializes tokens to a JSON file on disk.

InstalledFlow

MemoryStorage
A storage that remembers values for one session only.

NullStorage
A storage that remembers nothing.

PollInformation
Contains state of pending authentication requests

RefreshFlow
Implements the Outh2 Refresh Token Flow.

Scheme
A scheme for use in hyper::header::Authorization

ServiceAccountAccess
A token source (GetToken) yielding OAuth tokens for services that use ServiceAccount authorization. This token source caches token and automatically renews expired ones.

ServiceAccountKey
JSON schema of secret service account key. You can obtain the key from the Cloud Console at https://console.cloud.google.com/.

Token
Represents a token as returned by OAuth2 servers.

Enums

FlowType
All known authentication types, for suitable constants

InstalledFlowReturnMethod
cf. https://developers.google.com/identity/protocols/OAuth2InstalledApp#choosingredirecturi

PollError
Encapsulates all possible results of a poll_token(...) operation

RefreshResult
All possible outcomes of the refresh flow

Retry
A utility type to indicate how operations DeviceFlowHelper operations should be retried

TokenType
Represents all implemented token types

Constants

GOOGLE_DEVICE_CODE_URL

Traits

AuthenticatorDelegate
A partially implemented trait to interact with the Authenticator

GetToken
A provider for authorization tokens, yielding tokens valid for a given scope. The api_key() method is an alternative in case there are no scopes or if no user is involved.

TokenStorage
Implements a specialized storage to set and retrieve Token instances. The scope_hash represents the signature of the scopes for which the given token should be stored or retrieved. For completeness, the underlying, sorted scopes are provided as well. They might be useful for presentation to the user.

Functions

parse_application_secret
Read an application secret from a JSON string.

read_application_secret
Read an application secret from a file.

service_account_key_from_file
Read a service account key from a JSON file. You can download the JSON keys from the Google Cloud Console or the respective console of your service provider.