Trait yup_oauth2::authenticator_delegate::DeviceFlowDelegate

pub trait DeviceFlowDelegate: Send + Sync {
    fn present_user_code<'a>(
        &'a self,
        device_auth_resp: &'a DeviceAuthResponse
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'a>> { ... }
}

DeviceFlowDelegate methods are called when a device flow needs to ask the application what to do in certain cases.

Provided Methods

fn present_user_code<'a>(
&'a self,
device_auth_resp: &'a DeviceAuthResponse
) -> Pin<Box<dyn Future<Output = ()> + Send + 'a>>

The server has returned a user_code which must be shown to the user, along with the verification_uri.

Notes

Will be called exactly once, provided we didn’t abort during request_code phase.