Enum yup_oauth2::RefreshResult

pub enum RefreshResult {
    Error(Error),
    RefreshError(String, Option<String>),
    Success(Token),
}

All possible outcomes of the refresh flow

Variants

Error

Indicates connection failure

RefreshError

The server did not answer with a new token, providing the server message

Success

The refresh operation finished successfully, providing a new Token