Enum yup_oauth2::Retry

pub enum Retry {
    Abort,
    After(Duration),
    Skip,
}

A utility type to indicate how operations DeviceFlowHelper operations should be retried

Variants

Abort

Signal you don't want to retry

After

Signals you want to retry after the given duration

Skip

Instruct the caller to attempt to keep going, or choose an alternate path. If this is not supported, it will have the same effect as Abort