Trait yup_oauth2::GetToken

pub trait GetToken {
    fn token<'b, I, T>(&mut self, scopes: I) -> Result<Token, Box<dyn Error>>
    where
        T: AsRef<str> + Ord + 'b,
        I: IntoIterator<Item = &'b T>;

    fn api_key(&mut self) -> Option<String>;
}

A provider for authorization tokens, yielding tokens valid for a given scope. The api_key() method is an alternative in case there are no scopes or if no user is involved.

Required methods

fn token<'b, I, T>(&mut self, scopes: I) -> Result<Token, Box<dyn Error>> where
T: AsRef<str> + Ord + 'b,
I: IntoIterator<Item = &'b T>,

fn api_key(&mut self) -> Option<String>

Loading content...